llm/requirements.txt
torch==2.6.0
transformers[hf_xet]==4.51.0
hydra-core==1.3.2
omegaconf==2.3.0
sentencepiece==0.2.0
accelerate==1.6.0
huggingface_hub==0.30.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
